LottoGenerator - package 이동, 인스턴스 변수 분리 및 상수화 (로또 45개 번호 리스트), 테스트 추가 및 수정
