+++ src/test/java/lotto/LottoTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.util.LottoBasicNumber;
+import lotto.util.LottoGenerator;
 import org.junit.Test;
 
 import java.util.ArrayList;
@@ -11,8 +13,8 @@ public class LottoTest {
 
     @Test
     public void 일치하는_번호_개수_구하기() {
-        LottoGenerator lottoGenerator = new LottoGenerator(6);
-        Lotto lotto = new Lotto(lottoGenerator.generate());
+        LottoGenerator lottoGenerator = new LottoGenerator();
+        Lotto lotto = new Lotto(lottoGenerator.generate(LottoBasicNumber.numbers));
 
         List<Integer> numbers = new ArrayList<>();
         for (int i = 0; i < 6; i++) {
